backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
openai==1.6.1
langchain==0.0.340
langchain-openai==0.0.2
chromadb==0.4.18
pymilvus==2.3.4
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0
boto3==1.34.0
prometheus-client==0.19.0